Frontend refactor chore: move to typescript chore: use refreshContent for initialization (fixes #3, this is fixed actually) chore: use dom based initialization check (usually doesn't cause the regression of #11) fix: css build cross-platform fix: slide and gallery carousel syncing remove: duplicated style imported in javascript remove: duplicated carousel event registration
